+++ packages/core/src/events/RenderEditorIndicator.tsx
@@ -1,7 +1,7 @@
-import { RenderIndicator } from '@craftjs/utils';
 import React, { useEffect } from 'react';
 
 import { useEventHandler } from './EventContext';
+import { RenderIndicator } from './RenderIndicator';
 import movePlaceholder from './movePlaceholder';
 
 import { useInternalEditor } from '../editor/useInternalEditor';
