+++ packages/core/src/events/index.tsx
@@ -3,3 +3,4 @@ export * from './CoreEventHandlers';
 export * from './DefaultEventHandlers';
 export * from './EventHandlers';
 export * from './Events';
+export * from './RenderIndicator';
