+++ packages/layers/src/events/RenderLayerIndicator.tsx
@@ -1,5 +1,4 @@
-import { useEditor } from '@craftjs/core';
-import { RenderIndicator } from '@craftjs/utils';
+import { useEditor, RenderIndicator } from '@craftjs/core';
 import React, { useMemo } from 'react';
 
 import { useLayerManager } from '../manager/useLayerManager';
