+++ packages/utils/src/index.ts
@@ -1,6 +1,4 @@
 export * from './useMethods';
 export * from './useCollector';
-export * from './RenderIndicator';
 export * from './useEffectOnce';
 export * from './History';
-export * from './getRandomId';
